Refactor: Screens and navigation
- Refactor screens so that all screens have their own module
- Create `src/types` folder for types that belong to a generalized
  group of components and other kinds of modules.
- Create a new path rewrite for `src/types` as `_types`.
- Move app navigation type to `_types`. Update references to the
  module.
