+++ src/components/screens/Messages.screen.tsx
@@ -1,7 +1,7 @@
 import React, { type FC } from "react";
 import { type NativeStackScreenProps } from "@react-navigation/native-stack";
 import { SafeAreaView } from "react-native-safe-area-context";
-import { AppNavigatorParams } from "../navigators/App.navigator";
+import { type AppNavigatorParams } from "_types/navigation.types";
 
 import {
   Button,
