+++ src/types/navigation.types.ts
@@ -1,0 +1,8 @@
+export type AppNavigatorParams = {
+  Home: undefined;
+  Messages: undefined;
+  Login: undefined;
+  Help: {
+    content: string;
+  };
+};
