+++ tsconfig.json
@@ -8,7 +8,8 @@
     /* Completeness */
     "skipLibCheck": true,                                 /* Skip type checking all .d.ts files. */
     "paths": {
-      "_screens/*": ["src/components/screens/*"]
+      "_screens/*": ["src/components/screens/*"],
+      "_types/*": ["src/types/*"],
     }
   }
 }
